% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Jedem Tierchen sein Pläsierchen}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A character vector of length 1}

\item{sound}{A character vector of length 1}
}
\value{
A string
}
\description{
This little functions can help you to print the sound that an animal makes.
}
\examples{
animal_sounds("dog","woof")

}
